# Conflicts:
#	api-gateway-service/src/main/resources/application.yaml

+++ order-service/src/main/java/com/yandex/market/orderservice/model/ReceiptMethod.java
@@ -35,5 +35,5 @@ public class ReceiptMethod {
 
     private LocalTime deliveryEnd;
 
-    private double deliveryCost;
+    private Long deliveryCost;
 }
